GUI-R1/requirements.txt
accelerate
codetiming
datasets
liger-kernel
mathruler
numpy
omegaconf
pandas
peft
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils
ray[default]
tensordict
torchdata
transformers>=4.49.0
vllm>=0.7.3
wandb
